# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: OpsTower API
  description: |
    OpsTower is a comprehensive ridesharing platform operations API for the Philippine market.

    This API provides endpoints for:
    - Authentication & Authorization (RBAC with MFA)
    - Booking Management
    - Payment Processing (Maya, GCash)
    - Compliance Reporting (BSP, LTFRB, BIR)
    - Real-time Operations
    - Administrative Functions

    ## Authentication

    All API requests require authentication using JWT (JSON Web Token).
    Include the token in the Authorization header:

    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## Rate Limiting

    - **Standard**: 1000 requests per hour
    - **Authenticated**: 5000 requests per hour
    - **Admin**: Unlimited

    ## Error Codes

    See [Error Codes Reference](./ERROR_CODES.md) for complete list.

    ## Regions

    Philippine Timezone: Asia/Manila (UTC+8)

  version: 1.0.0
  contact:
    name: OpsTower API Support
    email: api@opstower.com
    url: https://docs.opstower.com
  license:
    name: Proprietary
    url: https://opstower.com/license
servers:
  - url: https://api.opstower.com
    description: Production server
  - url: https://staging-api.opstower.com
    description: Staging server
  - url: http://localhost:4000
    description: Local development server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from /api/auth/login endpoint.
        Format: Bearer <token>
  schemas:
...
    PaymentTransaction:
      type: object
      properties:
        id:
          type: string
          example: "txn_1234567890"
        transactionId:
          type: string
        amount:
          type: number
          format: float
          example: 250.00
        currency:
          type: string
          example: "PHP"
        status:
          type: string
          enum: [pending, processing, completed, failed, refunded]
        paymentMethod:
          type: string
          enum: [maya, gcash, cash, card]
        userId:
          type: string
        bookingId:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
    MayaPaymentRequest:
      type: object
      required:
        - amount
        - description
        - userId
        - userType
        - customerName
        - customerEmail
        - successUrl
        - failureUrl
      properties:
        amount:
          type: number
          format: float
          minimum: 1
          example: 250.00
        description:
          type: string
          example: "Ride payment for booking #123"
        userId:
          type: string
        userType:
          type: string
          enum: [passenger, driver, operator]
        customerName:
          type: string
          example: "Juan dela Cruz"
        customerEmail:
          type: string
          format: email
          example: "juan@example.com"
        customerPhone:
          type: string
          example: "+639171234567"
        bookingId:
          type: string
          nullable: true
        successUrl:
          type: string
          format: uri
        failureUrl:
          type: string
          format: uri
        cancelUrl:
          type: string
          format: uri
        metadata:
          type: object
          additionalProperties: true
    MayaPaymentResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            transactionId:
              type: string
            referenceNumber:
              type: string
            checkoutId:
              type: string
            redirectUrl:
              type: string
              format: uri
            amount:
              type: number
              format: float
            currency:
              type: string
              example: "PHP"
            expiresAt:
              type: string
              format: date-time
            status:
              type: string
              example: "pending"

    # RBAC Schemas
...
paths:
...
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                      refreshToken:
                        type: string
...
  /api/payments/maya/initiate:
    post:
      tags:
        - Maya
        - Payments
      summary: Initiate Maya payment
      description: Start a new Maya payment transaction
      operationId: initiateMayaPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MayaPaymentRequest'
      responses:
        '200':
          description: Payment initiated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MayaPaymentResponse'
        '400':
          description: Invalid payment request
  /api/payments/maya/status/{transactionId}:
    get:
      tags:
        - Maya
        - Payments
      summary: Get Maya payment status
      operationId: getMayaPaymentStatus
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/PaymentTransaction'
  /api/payments/maya/webhook:
    post:
      tags:
        - Maya
        - Payments
      summary: Maya webhook
      description: Webhook endpoint for Maya payment notifications
      operationId: mayaWebhook
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed
  /api/payments/maya/refund:
    post:
      tags:
        - Maya
        - Payments
      summary: Refund Maya payment
      operationId: refundMayaPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - transactionId
                - amount
              properties:
                transactionId:
                  type: string
                amount:
                  type: number
                  format: float
                reason:
                  type: string
      responses:
        '200':
          description: Refund initiated successfully

  # ==========================================================================
  # PAYMENT ENDPOINTS - GCASH
  # ==========================================================================
...
  /api/payments/gcash/status/{transactionId}:
    get:
      tags:
        - GCash
        - Payments
      summary: Get GCash payment status
      operationId: getGCashPaymentStatus
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment status retrieved
...